Decentraland Games Odds Calculator
Quick script to determine the odds on the casino floor of the decentraland decentral games casino's.  In the casino, you can play with MANA, DAI, ATRI, and ETH, and for each bet, you are rewarded in DG.  This means price fluctuation between currencies change the odds of the game.  With this script, you can see the current odds of each game, with varying levels of rewards bonuses activated.  It utilizes the cryptocompare API and runs in terminal.